renaming some components to reflect System Map
