+++ firefly/src/styles/TabletLandingPage.scss
